+++ queries/tsx/rainbow-tags.scm
@@ -3,3 +3,7 @@
 (jsx_element
   open_tag: (jsx_opening_element) @opening
   close_tag: (jsx_closing_element) @closing) @container
+
+(jsx_expression
+  "{" @opening
+  "}" @closing) @container