+++ test/highlight/jsx/regular.jsx
@@ -31,6 +31,9 @@ function app() {
 			<p>
 				This is an <a href="https://example.com">Example<br/>link</a> with<br/> line <br/>break.
 			</p>
+			<ComponentWithChildren>
+				{someFunction().map((x) => <div></div>)}
+			</ComponentWithChildren>
 			<button onClick={hello}>Click me!</button>
 		</div>
 	)