+++ test/highlight/tsx/regular.tsx
@@ -66,6 +66,9 @@ function app() {
 				This is an <a href="https://example.com">Example<br/>link</a> with<br/> line <br/>break.
 			</p>
 			<button onClick={hello}>Click me!</button>
+			<ComponentWithChildren>
+				{someFunction().map((x) => <div></div>)}
+			</ComponentWithChildren>
 			<CustomComponent bool={true} arr={[1, 2, 3]} />
 		</div>
 	)