+++ CHANGELOG.rst
@@ -22,7 +22,7 @@ Added
 Fixed
 =====
 
-- Missing patterns in Typescript React query `rainbow-tags`
+- Missing patterns in React queries
 - Obsolete extmarks being preserved when text changes region (e.g. in Markdown
   moving a line out of a Lua code block into the surrounding text body)